Move `setContext` into `setupContext`.
