+++ addon-test-support/index.js
@@ -3,10 +3,9 @@ export { default as TestModule } from './legacy-0-6-x/test-module';
 export { default as TestModuleForAcceptance } from './legacy-0-6-x/test-module-for-acceptance';
 export { default as TestModuleForComponent } from './legacy-0-6-x/test-module-for-component';
 export { default as TestModuleForModel } from './legacy-0-6-x/test-module-for-model';
-export { getContext, setContext, unsetContext } from './legacy-0-6-x/test-context';
 
 export { setResolver } from './resolver';
-export { default as setupContext } from './setup-context';
+export { default as setupContext, getContext, setContext, unsetContext } from './setup-context';
 export { default as teardownContext } from './teardown-context';
 export { default as setupRenderingContext } from './setup-rendering-context';
 export { default as teardownRenderingContext } from './teardown-rendering-context';
